Validations to services

+++ src/services/DeleteTransactionService.ts
@@ -1,7 +1,6 @@
-// import AppError from '../errors/AppError';
-
-import { getRepository } from 'typeorm';
-import Transaction from '../models/Transaction';
+import { getCustomRepository } from 'typeorm';
+import AppError from '../errors/AppError';
+import TransactionsRepository from '../repositories/TransactionsRepository';
 
 interface DeleteTransactionDTO {
   id: string;
@@ -9,7 +8,20 @@ interface DeleteTransactionDTO {
 
 class DeleteTransactionService {
   public async execute({ id }: DeleteTransactionDTO): Promise<void> {
-    const repository = getRepository(Transaction);
+    const repository = getCustomRepository(TransactionsRepository);
+
+    const transaction = await repository.findOne(id);
+    if (!transaction) {
+      throw new AppError('Invalid transaction id', 400);
+    }
+
+    const { total } = await repository.getBalance();
+    if (transaction.type === 'income' && total < transaction.value) {
+      throw new AppError(
+        "You don't have enough balance to remove this transaction",
+        400,
+      );
+    }
 
     await repository.delete(id);
   }
